photorename/requirements.txt
pandas
pillow
pillow-heif
opencv-contrib-python
pytesseract
